buildbox_challenge_final/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
httpx==0.24.1
python-dotenv==1.0.0
sqlalchemy==1.4.49
pydantic==1.10.7
